Move images and create an image collection
