+++ frontend/src/assets/images/bunny_avatar.png
