+++ frontend/src/assets/images/index.jsx
@@ -1,0 +1,27 @@
+import AlmondBreeze from "./almond-breeze.png";
+import Bread from "./bread.png";
+import Dairy from "./dairy.png";
+import Farmer from "./farmer.png";
+import Fish from "./fish.png";
+import Fruit from "./fruit-veg.png";
+import GoogleStoreBadge from "./Google_Play_Store_badge.png";
+import Logo from "./Logo.png";
+import Meat from "./meat.png";
+import Milk from "./milk.png";
+import Wheat from "./wheat.png";
+
+const Images = {
+  AlmondBreeze,
+  Bread,
+  Dairy,
+  Farmer,
+  Fish,
+  Fruit,
+  GoogleStoreBadge,
+  Logo,
+  Meat,
+  Milk,
+  Wheat,
+};
+
+export default Images;
